+++ src/raf.c
@@ -68,10 +68,10 @@ int main(int argc, char **argv) {
 
     // Keep track of the number of sites at which the number
     // of reference or alternate alleles differs from 0
-    int         zeroref = 0, zeroalt = 0, zerogtype = 0;
-    int         missref = 0;
-    int         multref = 0, multalt = 0;
-    int         nbad = 0, ngood = 0;
+    long int         zeroref = 0, zeroalt = 0, zerogtype = 0;
+    long int         missref = 0;
+    long int         multref = 0, multalt = 0;
+    long int         nbad = 0, ngood = 0;
     int         ok;             // is current line acceptable
     long unsigned lastnucpos = 0, nucpos;
     char        lastchr[100] = { '\0' };
@@ -259,21 +259,21 @@ int main(int argc, char **argv) {
         printf("%s\t%s\t%s\t%s\t%0.18g\n",
                chr, pos, ref[0], alt[0], p);
     }
-    fprintf(stderr, "raf: %d good sites; %d rejected\n", ngood, nbad);
+    fprintf(stderr, "raf: %ld good sites; %ld rejected\n", ngood, nbad);
     if(zeroref)
-        fprintf(stderr, "raf: bad sites with 0 ref alleles: %d\n", zeroref);
+        fprintf(stderr, "raf: bad sites with 0 ref alleles: %ld\n", zeroref);
     if(zeroalt)
-        fprintf(stderr, "raf: bad sites with 0 alt alleles: %d\n", zeroalt);
+        fprintf(stderr, "raf: bad sites with 0 alt alleles: %ld\n", zeroalt);
     if(zerogtype)
-        fprintf(stderr, "raf: bad sites with 0 genotypes: %d\n", zerogtype);
+        fprintf(stderr, "raf: bad sites with 0 genotypes: %ld\n", zerogtype);
     if(multref)
-        fprintf(stderr, "raf: bad sites with multiple ref alleles: %d\n",
+        fprintf(stderr, "raf: bad sites with multiple ref alleles: %ld\n",
                 multref);
     if(multalt)
-        fprintf(stderr, "raf: bad sites with multiple alt alleles: %d\n",
+        fprintf(stderr, "raf: bad sites with multiple alt alleles: %ld\n",
                 multalt);
     if(missref)
-        fprintf(stderr, "raf: bad sites with missing ref alleles: %d\n",
+        fprintf(stderr, "raf: bad sites with missing ref alleles: %ld\n",
                 missref);
     return 0;
 }
